% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_agg2daily.R
\name{agg2daily}
\alias{agg2daily}
\title{Aggregates a data frame to daily resolution}
\usage{
agg2daily(df, col_name, fun, allow_na = 0)
}
\arguments{
\item{df}{data frame with class Date or POSIXct in the first column.}

\item{col_name}{string (vector) with the name(s) of the column(s) to aggregate.}

\item{fun}{string (vector) containing the name(s) of function(s) to use: mean, min, max, sum, last or first.}

\item{allow_na}{optional. Numeric value with the maximum allowed number of NA_real_ values. By default the function will not tolerate any NA_real_ in an aggregation period (and will return NA_real_ instead).}
}
\value{
A data frame with the Date and the aggregated variable(s).
}
\description{
Aggregates a data frame to daily resolution
}
\keyword{internal}
