% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_build_table.R
\name{build_table}
\alias{build_table}
\title{Create a basic data frame for plotting}
\usage{
build_table(hm_obj, slot_name, col_name, from = NULL, to = NULL)
}
\arguments{
\item{hm_obj}{a valid \code{hydromet_xxx} class object.}

\item{slot_name}{string vector with the slot(s) to extract from the \code{hm_obj}.}

\item{col_name}{a list containing a vector string per slot with the name(s) of the column(s) to plot.}

\item{from}{string (or \code{POSIXct}) with the starting \code{date}. You can use 'from' without 'to'. In this case you will subset your data 'from' till the end.}

\item{to}{string (or \code{POSIXct}) with the ending \code{date}. You can use 'to' without 'from'. In this case you will subset your data from the beginning till 'to'.}
}
\value{
A data frame with the date in the first column and the variables to plot on the others.
}
\description{
Create a basic data frame for plotting
}
\keyword{internal}
