% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Check class for arguments}
\usage{
check_class(argument, target, arg_name)
}
\arguments{
\item{argument}{the argument to evaluate.}

\item{target}{string with the allowed class.}

\item{arg_name}{string with argument name to print (just for error message)}
}
\value{
An error message if \option{argument} does not fit the \option{target} requirements
}
\description{
Check class for arguments
}
\keyword{internal}
