% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cross.R
\name{check_cross}
\alias{check_cross}
\title{Check cross for argument coherence}
\usage{
check_cross(ref_arg, eval_arg, arg_names)
}
\arguments{
\item{ref_arg}{the argument for reference length.}

\item{eval_arg}{the argument to evaluate.}

\item{arg_names}{vector string with arguments names (\code{ref_arg} and \code{eval_arg})}
}
\value{
An error message if \option{eval_arg} does not fit the \option{ref_arg} length.
}
\description{
Check cross for argument coherence
}
\keyword{internal}
