% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{check_length}
\alias{check_length}
\title{Check argument length}
\usage{
check_length(argument, max_allow, arg_name)
}
\arguments{
\item{argument}{the argument to evaluate.}

\item{max_allow}{maximum argument length}

\item{arg_name}{string with argument name to print (just for error message)}
}
\value{
An error message if \option{argument} does not fit the \option{max_allow} requirements
}
\description{
Check argument length
}
\keyword{internal}
