% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric argument consistency}
\usage{
check_numeric(argument, target, arg_name)
}
\arguments{
\item{argument}{the argument to evaluate.}

\item{target}{allowed numeric values.}

\item{arg_name}{string with argument name to print (just for error message)}
}
\value{
An error message if \option{argument} does not fit the \option{target} requirements
}
\description{
Check numeric argument consistency
}
\keyword{internal}
