% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_string.R
\name{check_string}
\alias{check_string}
\title{Check string argument consistency}
\usage{
check_string(argument, target, arg_name)
}
\arguments{
\item{argument}{the argument to evaluate.}

\item{target}{allowed string values.}

\item{arg_name}{string with argument name to print (just for error message)}
}
\value{
An error message if \option{argument} does not fit the \option{target} requirements
}
\description{
Check string argument consistency
}
\keyword{internal}
