% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_col_stats.R
\name{col_stats}
\alias{col_stats}
\title{Get an specific matrix columns statistics.}
\usage{
col_stats(x, stats, allow_na)
}
\arguments{
\item{x}{numeric matrix}

\item{stats}{string vector with the statistics to compute (sum, max, min, mean, first or last)}

\item{allow_na}{number of allowed NA_real_}
}
\value{
Numeric vector with one value per matrix column
}
\description{
Get an specific matrix columns statistics.
}
\keyword{internal}
