% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_dual_y_table.R
\name{dual_y_table}
\alias{dual_y_table}
\title{Rescale the data frame values in order to graph a dual y axis plot}
\usage{
dual_y_table(df, y_left, y_right)
}
\arguments{
\item{df}{data frame provided by \code{build_table} function.}

\item{y_left}{string vector with column names of the variables to plot in the left y axis.}

\item{y_right}{string vector with column names of the variables to plot in the right y axis.}
}
\value{
A list containing the rescaled data frame and the transformation formula.
}
\description{
Rescale the data frame values in order to graph a dual y axis plot
}
\keyword{internal}
