% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hm_create.R
\name{hm_create}
\alias{hm_create}
\title{Creates an hydromet object.}
\usage{
hm_create(class_name = "station")
}
\arguments{
\item{class_name}{string with the name of the class. Valid arguments are: \code{hydromet}, \code{station} or \code{compact}.}
}
\value{
An S4 object of class \code{hydromet}.
}
\description{
This function is the constructor of \code{hydromet} class and its subclass.
}
\examples{
# create class 'hydromet'
hym_metadata <- hm_create(class_name = 'hydromet')

# subclass 'station'
hym_station <- hm_create(class_name = 'station')

# subclass 'compact'
hym_compact <- hm_create(class_name = 'compact')

}
